TrafficmonitoringAutoSystem/traffic_monitoring_app/traffic_monitoring_app/requirements.txt
fastapi==0.78.0
numpy==1.22.4
opencv_python==4.6.0.66
PyYAML==6.0
uvicorn==0.17.6
